requirements.txt
# ── Django core ───────────────────────────────────────────────────────────────
Django>=5.0,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# ── Images ────────────────────────────────────────────────────────────────────
Pillow>=10.0

# ── Payment ───────────────────────────────────────────────────────────────────
razorpay>=1.4

# ── HTTP client ───────────────────────────────────────────────────────────────
requests>=2.31

# ── Config ────────────────────────────────────────────────────────────────────
python-decouple>=3.8

# ── FCM Push Notifications (HTTP v1 API) ──────────────────────────────────────
google-auth>=2.23
google-auth-httplib2>=0.2

# ── Production server ─────────────────────────────────────────────────────────
gunicorn>=21.0

# ── Static files ──────────────────────────────────────────────────────────────
whitenoise>=6.6

# NOTE: No Celery, no Redis, no psycopg2 — not needed for PythonAnywhere + SQLite
# Async tasks are handled via Python's built-in threading module (see core/thread_tasks.py)
